Refactor ToDoItem, Change /api/list-obj to work with Hibernate
