+++ server/src/main/java/org.simple/entity/ToDoItemRepository.java
@@ -1,0 +1,10 @@
+package org.simple.entity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ToDoItemRepository extends JpaRepository<ToDoItem, Long> {
+
+    ToDoItem findById(Integer id);
+}
