+++ server/src/main/java/org.simple/util/ToDoItemUtils.java
@@ -1,0 +1,19 @@
+package org.simple.util;
+
+import org.simple.entity.ToDoItem;
+import org.simple.entity.ToDoItemRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+public class ToDoItemUtils {
+
+    @Autowired
+    private ToDoItemRepository toDoItemRepository;
+
+    public List<ToDoItem> getToDoList(){
+        return this.toDoItemRepository.findAll();
+    }
+}
