+++ server/src/test/resources/application.properties
@@ -6,3 +6,9 @@ spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.datasource.url=jdbc:h2:mem:testdb
 
 spring.resources.static-locations=file:../client/build
+
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.H2Dialect
